requirements.txt
apache-beam==2.61.0
attrs==24.3.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
cloudpickle==2.2.1
crcmod==1.7
Deprecated==1.2.15
dill==0.3.1.1
dnspython==2.7.0
docopt==0.6.2
fastavro==1.10.0
fasteners==0.19
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-core==2.4.1
google-cloud-pubsub==2.27.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.13.1
grpcio==1.68.1
grpcio-status==1.68.1
hdfs==2.7.3
httplib2==0.22.0
idna==3.10
importlib_metadata==8.5.0
jsonpickle==3.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
numpy==2.1.3
objsize==0.7.0
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
orjson==3.10.12
packaging==24.2
proto-plus==1.25.0
protobuf==5.29.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydot==1.4.2
pymongo==4.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rpds-py==0.22.3
rsa==4.9
six==1.17.0
sortedcontainers==2.4.0
typing_extensions==4.12.2
urllib3==2.2.3
wrapt==1.17.0
zipp==3.21.0
zstandard==0.23.0
